backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
mlflow
pandas
numpy
groq
